+++ src/agentsec/threatmodel/__init__.py
@@ -1,11 +1,13 @@
 """STRIDE-for-agents threat modeling."""
 
+from agentsec.threatmodel.crosswalk import crosswalk_markdown
 from agentsec.threatmodel.render import (
     COMPONENT_TYPES,
     SYSTEM_TEMPLATE,
     SystemSpecError,
     Threat,
     load_catalog,
+    load_frameworks,
     load_system,
     render_markdown,
     validate_system,
@@ -16,7 +18,9 @@ __all__ = [
     "SYSTEM_TEMPLATE",
     "SystemSpecError",
     "Threat",
+    "crosswalk_markdown",
     "load_catalog",
+    "load_frameworks",
     "load_system",
     "render_markdown",
     "validate_system",